Added swap functions to wrapper. Updated example.
